Fixed sorting issue + added etherscan filters

+++ database/migrations/2022_07_03_063808_create_opensea_transactions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('wallet');
 
             $table->enum('schema', config('hawk.opensea.event.schema'));
-            $table->unsignedInteger('event_id');
+            $table->unsignedBigInteger('event_id');
             $table->enum('event_type', config('hawk.opensea.event.types'));
             $table->unsignedInteger('event_timestamp');
